Add static hostnames in generated profiles
We saw an issue where kubelet was starting before a hostname could be set via
dhcp offering, which resulted in nodes registering to the cluster as
`localhost`. Since we use explicit matchbox profiles per instance we can remedy
that via static hostnames so that nodes do not init as localhost anymore.