Exp 1 ld7 move (#36)
* Optional local disk partition for cfssl data

Adds a flag to allow partitioning space in the root disk that will
not be wiped during boot to store cfssl data. Defaults to false,
which will be noop for existing deployments.
To be used for env migration where we won't have access to iscsi
servers.

* Output local cfssl volume id

* Fix cfssl disk size

* Add option to create a partiton for persistent storage on workers

This will partition our disks and avoid wiping the space dedicated to
storage between boots. The default mount location is based on the one
expected from democratic-csi driver.

* Prefix locals to avoid conflicts in the future